Makes code work with an agnostic db approach (#1)
* Adds sqlmock, gorm, gotenv dependencies

* Makes the DB code 'db agnostic'
By using gorm the code is not tied to a specific
database and can be easily swapped for any DB as
required.

* Ignore env files

* Removes tailscale specific code
The original code had (for obvious reasons) code
that was relying on being on a tailscale network.
This is not my use case for now and I have
resorted to comment it out for now. In the future
the code should detect if it's running in a
tailscale net and use that code path or a more
generic one otherwise.

* Removes unused code
The previous code was relying always on a SQLite
DB which could have "snapshots" from a file. With
the DB-agnostic approach this is not the proper
way to restore a 'snapshot' and therefore it is
removed from the codebase

* Adds VSCode dev container configuration

* Modifies README to mention the changes